pyproject.toml
[project]
name = "ocr-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.27.2",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pypdf>=6.9.1",
]
